Create Designing a Store Database
This project involved the creation of a store database in SQL, consisting of a table with 5 columns and 15 rows. Data analysis was performed to find the minimum and maximum prices, determine available stock, sort items by price, and identify the count of items for each clothing brand using the group by function.