Create Marvel Character Analytics
In this SQL project, Marvel character data was utilized to uncover insightful information. The project involved calculating the average height for both male and female characters and identifying the top 10 most popular heroes. Additionally, the analysis found all male heroes who surpass the average height. To enhance the data, a new column was created in the table to quantify the strength level of each character. Through these SQL queries, valuable insights and comparisons were made to gain a deeper understanding of the Marvel character universe.